0.0.1

+++ src/Module_4RELAY.cpp
@@ -1,5 +1,6 @@
-#include "MODULE_4RELAY.h"
+#include "Module_4RELAY.h"
 
+/*! @brief Initialize the 4RELAY.*/
 bool MODULE_4RELAY::begin(TwoWire* wire, uint8_t addr, uint8_t sda, uint8_t scl,
                           uint32_t speed) {
     _wire = wire;
@@ -17,6 +18,7 @@ bool MODULE_4RELAY::begin(TwoWire* wire, uint8_t addr, uint8_t sda, uint8_t scl,
     }
 }
 
+/*! @brief Write data to the specified register address. */
 bool MODULE_4RELAY::writeReg(uint8_t data) {
     _wire->beginTransmission(_addr);
     _wire->write(MODULE_4RELAY_REG);
@@ -25,23 +27,27 @@ bool MODULE_4RELAY::writeReg(uint8_t data) {
     return false;
 }
 
+/*! @brief Read data to the specified register address. */
 uint8_t MODULE_4RELAY::readReg() {
     _wire->beginTransmission(_addr);
     _wire->write(MODULE_4RELAY_REG);
     _wire->endTransmission(false);
-    _wire->requestFrom(_addr, 1);
+    _wire->requestFrom(_addr, (uint8_t)1);
     return _wire->read();
 }
 
+/*! @brief Read all relay status. */
 uint8_t MODULE_4RELAY::getAllRelayState() {
     return readReg();
 }
 
+/*! @brief Read the status of the specified relay. */
 bool MODULE_4RELAY::getRelayState(uint8_t index) {
     uint8_t stateByte = getAllRelayState();
     return (stateByte >> index) & 0x01;
 }
 
+/*! @brief Set the state of the specified relay. */
 bool MODULE_4RELAY::setRelay(uint8_t index, bool state) {
     uint8_t stateByte = getAllRelayState();
     if (state) {
@@ -52,6 +58,7 @@ bool MODULE_4RELAY::setRelay(uint8_t index, bool state) {
     return writeReg(stateByte);
 }
 
+/*! @brief Set all relay states. */
 bool MODULE_4RELAY::setAllRelay(bool state) {
     return writeReg(state ? 0xff : 0x00);
 }
